refactor: MoveContext 및 궁성 관련 도메인 패키지 구조 개선
- MoveContext를 domain.move 패키지로 분리
- Palace, PalaceArea, PalaceRoute를 palace 패키지로 이동
- 기물에서 궁성/이동 관련 로직을 MoveContext로 위임
- General, Guard 등 중복된 이동 로직 제거 및 단순화
- Soldier의 궁성 이동 조건을 MoveContext 기반으로 개선
