+++ src/main/java/domain/piece/Chariot.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.board.MoveContext;
+import domain.move.MoveContext;
 import domain.coordination.Coordination;
 import domain.piece.error.PieceException;
 import java.util.List;
